Separate Wasm from other end-to-end tests. (#1013)
* Separate Wasm from other end-to-end tests.

This removes a lot of conditional compilation based on the Wasm flags.

* Remove some more conditional compilation.
