+++ linera-service/tests/common/mod.rs
@@ -1,0 +1,8 @@
+// Copyright (c) Zefchain Labs, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+use once_cell::sync::Lazy;
+use tokio::sync::Mutex;
+
+/// A static lock to prevent integration tests from running in parallel.
+pub static INTEGRATION_TEST_GUARD: Lazy<Mutex<()>> = Lazy::new(|| Mutex::new(()));
